refactored app. setup functions in their own script and remove hello world home page
