+++ .gitignore
@@ -1,2 +1,3 @@
 .devdbrc
 *.db
+*.pyc
