+++ templates/new_recipe.html
@@ -25,6 +25,6 @@
 </form>
 
 <br>
-<a href="/recipes">View Recipes</a>
+<a href="/">View Recipes</a>
 </body>
 </html>
