+++ templates/recipes.html
@@ -6,7 +6,7 @@
 <body>
     <h1>Recipes</h1>
 
-    <form method="GET" action="/recipes">
+    <form method="GET" action="/">
         <label for = "id">Choose a recipe:</label>
         <select name = "id" id = "id">
             {% for recipe in recipes %}
